+++ CHANGELOG.md
@@ -20,6 +20,8 @@
   now this is fixed and auto-reset-formatting works as expected
 * added a new param <code>ignore_in_strings: *bool* = True</code> to `Regex.brackets()`:<br>
   if this param is true and a bracket is inside a string (e.g. `'...'` or `"..."`), it will not be counted as the matching closing bracket
+* removed `lru_cache` from the `xx_format_codes` module's internal methods, since it was unnecessary
+* adjusted `FormatCodes.__config_console()` so it can only be called once per process
 
 ## 20.01.2025 `v1.6.2`
 * moved the method `is_admin()` from `xx_console` to `xx_system`